Refactor: Update architecture, move runtime datasets to data/, and implement dashboard features
