--- Locations.csv
@@ -1,25 +1,0 @@
-latitude,longitude,location_name
-19.0734,72.9972,"Vashi, Navi Mumbai, Thane Taluka, Thane, Maharashtra, 400703, India"
-19.159,72.9986,"Airoli Station Road, Airoli Sector 2, Airoli, Navi Mumbai, Thane Taluka, Thane, Maharashtra, 400708, India"
-19.033,73.0297,"Amra Marg, Seawoods East, Navi Mumbai, Thane Taluka, Thane, Maharashtra, 400614, India"
-18.885,72.95,"Uran, Uran Taluka, Raigad, Maharashtra, 400702, India"
-19.0335,73.12,"Panvel, Panvel Taluka, Raigad, Maharashtra, 401218, India"
-19.1213,73.1395,"Panvel, Panvel Taluka, Raigad, Maharashtra, 410208, India"
-19.1659,73.0726,"1, Casa Bella Road, Palava Phase 1, Nilaje, Kalyan-Dombivli, Kalyan Taluka, Thane, Maharashtra, 421204, India"
-18.9883,73.0652,"Navi Mumbai International Airport, Amra Marg, Moha Village, Targhar, Navi Mumbai, Panvel Taluka, Raigad, Maharashtra, 400614, India"
-19.1357,73.1962,"Shelter with food, Malang Machi Route, Haji Malangwadi, Ambernath Taluka, Thane, Maharashtra, 400010, India"
-19.1733,72.9495,"Kalidas Sports Complex, Purushottam Kheraj Marg, Mirani Nagar, T Ward, Zone 6, Mumbai, Maharashtra, 400080, India"
-19.1251,72.8845,"Marol Police Camp, Marol Maroshi Village Road, Vijay Nagar, K/E Ward, Zone 3, Mumbai, Maharashtra, 400059, India"
-19.0632,72.9024,"P.L. Lokhande Marg, Adarsh Nagar, M/W Ward, Zone 5, Mumbai, Maharashtra, 400089, India"
-18.9919,73.1265,"Jhanardan A Bhagat Marg, New Panvel, Panvel, Panvel Taluka, Raigad, Maharashtra, 497586, India"
-19.0197,72.8475,"Scheme 5 Road no.3, Dadar Hindu Colony, F/N Ward, Zone 2, Mumbai, Maharashtra, 400014, India"
-18.9254,73.1437,"Bhatan, Panvel Taluka, Raigad, Maharashtra, 410221, India"
-19.0423,73.171,"Khanav, Panvel Taluka, Raigad, Maharashtra, 400010, India"
-18.96127,72.986389,"Nhava Khadi, Panvel Taluka, Raigad, Maharashtra, India"
-18.95806,72.80318,"Godrej Baug Road, Malabar Hill, D Ward, Zone 1, Mumbai, Maharashtra, 400006, India"
-19.12591,72.82054,"Seven Bungalows, K/W Ward, Zone 3, Mumbai, Maharashtra, 400061, India"
-19.1015,72.82679,"JW Marriott, Juhu Tara Road, Juhu, K/W Ward, Zone 3, Mumbai, Maharashtra, 400049, India"
-19.19714,73.2424,"Kalyan Taluka, Thane, Maharashtra, 421503, India"
-19.19925,73.02163,"Mumbra, Reti Bandar, Thane, Thane Taluka, Thane, Maharashtra, 400605, India"
-19.19695,72.98242,"Kranti Nagar, Jambli Naka, Thane, Thane Taluka, Thane, Maharashtra, 400603, India"
-18.99679,72.81688,"Dr Annie Besant Road, Mariammar Nagar, G/S Ward, Zone 2, Mumbai, Maharashtra, 400018, India"
