+++ data/Locations.csv
@@ -1,0 +1,25 @@
+latitude,longitude,location_name
+19.0734,72.9972,"Vashi, Navi Mumbai"
+19.159,72.9986,"Airoli, Navi Mumbai"
+19.033,73.0297,"Seawoods, Navi Mumbai"
+18.885,72.95,"Uran, Raigad"
+19.0335,73.12,"Kalamboli, Panvel"
+19.1213,73.1395,"Wadi, Panvel"
+19.1659,73.0726,"Nilje, Kalyan"
+18.9883,73.0652,"Navi Mumbai International Airport, Navi Mumbai"
+19.1357,73.1962,"Badlapur"
+19.1733,72.9495,"Mulund, Mumbai"
+19.1251,72.8845,"Powai, Mumbai"
+19.0632,72.9024,"Chembur, Mumbai"
+18.9919,73.1265,"New Panvel"
+19.0197,72.8475,"Dadar, Mumbai"
+18.9254,73.1437,"Shirdhon, Panvel"
+19.0423,73.171,"Khanav, Panvel"
+18.96127,72.986389,"Nhava Khadi, Panvel"
+18.95806,72.80318,"Malabar Hill, Mumbai"
+19.12591,72.82054,"Versova, Mumbai"
+19.1015,72.82679,"Juhu, Mumbai"
+19.19714,73.2424,"Ambernath"
+19.19925,73.02163,"Mumbra, Thane"
+19.19695,72.98242,"Thane City"
+18.99679,72.81688,"Worli, Mumbai"
