+++ data/water_rep_data.csv
@@ -1,0 +1,18 @@
+location_name,lat,lon,pH,dissolvedoxygen,bod,cod,nitrate,FecalColiform,representative_wqi
+Worli Seaface,18.9975,72.8089,8.0,5.7,10,96,0.74,540,
+Vashi Creek,19.0755,72.9837,7.5,4.3,12,100,1.19,110,
+Panvel Creek,19.0042,73.0327,7.3,5.9,7,48,2.8,34,
+"Thane Creek, Kalwa Bridge",19.1969,72.9849,7.5,4.5,13,156,1.5,150,
+"Ulhas River, Badlapur",19.139,73.2133,7.4,6.8,2.4,8,0.45,7.8,
+"Sea Water, Gateway",18.9215,72.8337,8.0,4.0,13,84,0.72,540,
+"Mahim creek, Bandra",19.026,72.8338,8.0,5.3,10,92,0.72,150,
+"Bhatsa River, Pise dam",19.3145,73.184,7.6,7.3,2.8,12,0.36,9.3,
+"Vashi Creek, Airoli",19.1601,72.9817,7.4,4.8,10,88,1.16,79,
+"Thane Creek, Elephanta",18.9641,72.9428,7.9,4.2,9,108,0.8,120,
+Malabar Hills,18.9504,72.8038,8.0,4.4,12,104,0.72,350,
+"Versova  beach, andheri",19.1239,72.8168,7.7,3.4,14,240,0.86,280,
+Juhu beach,19.1013,72.8233,7.5,4.2,15,272,0.86,920,
+"Ulhas river, Mohane, Kalyan",19.2559,73.1789,7.7,7.2,2.2,8,0.53,11,
+"Ulhas river, Jambhul, Ambernath",19.1994,73.2467,7.5,6.9,2.4,8,0.47,7.8,
+"Ulhas creek, Mumbra ",19.2013,73.0237,7.6,4.3,12,132,2.61,46,
+"Bassein creek, Vasai",19.3195,72.8459,7.6,3.6,15,200,3.48,31,
