+++ static/css/dashboard.css
@@ -38,8 +38,7 @@ header {
     display: flex; 
     align-items: center; 
     gap: 16px;
-    position: sticky;
-    top: 0;
+    position: relative;
     z-index: 100;
 }
 
@@ -129,29 +128,35 @@ header h1 {
     flex-wrap: wrap; 
 }
 
-select { 
-    background: var(--surface);
-    backdrop-filter: blur(8px);
-    color: var(--text); 
-    border: 1px solid var(--card-border);
-    border-radius: 12px; 
-    padding: 12px 16px; 
-    font-size: 14px; 
-    font-weight: 500;
+select {
+    background: #1e293b;
+    color: var(--text);
+    border: 1px solid #334155;
+    border-radius: 8px;
+    padding: 10px 14px;
+    padding-right: 36px;
+    font-size: 14px;
     cursor: pointer;
-    transition: all 0.2s ease;
+    width: 100%;
+    max-width: 400px;
     outline: none;
-    min-width: 200px;
+    appearance: none;
+    background-image: url("data:image/svg+xml;charset=US-ASCII,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%2214%22%20height%3D%228%22%20viewBox%3D%220%200%2014%208%22%20fill%3D%22none%22%3E%3Cpath%20d%3D%22M1%201L7%207L13%201%22%20stroke%3D%22%239CA3AF%22%20stroke-width%3D%222%22%20stroke-linecap%3D%22round%22%20stroke-linejoin%3D%22round%22%2F%3E%3C%2Fsvg%3E");
+    background-repeat: no-repeat;
+    background-position: right 14px center;
+    background-size: 12px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    transition: all 0.2s ease;
 }
 
 select:hover {
-    border-color: var(--accent);
-    background: rgba(56, 189, 248, 0.05);
+    border-color: #475569;
+    background-color: #0f172a;
 }
 
 select:focus {
     border-color: var(--accent);
-    box-shadow: 0 0 0 3px var(--accent-glow);
+    box-shadow: 0 0 0 3px rgba(56, 189, 248, 0.2);
 }
 
 .legend {
