--- water_rep_data.csv
@@ -1,18 +1,0 @@
-location_name,lat,lon,pH,dissolvedoxygen,bod,cod,nitrate,FecalColiform,representative_wqi
-Worli Seaface,18.9975,72.8089,7.9,5.4,9,64,1.44,110,62.19
-Vashi Creek,19.0755,72.9837,7.7,5.6,8,40,1.89,140,65.17
-Panvel Creek,19.0042,73.0327,7.6,6.8,3.2,8,0.9,63,78.28
-"Thane Creek, Kalwa Bridge",19.1969,72.9849,7.7,6.4,4.8,20,1.17,94,75.18
-"Ulhas River, Badlapur",19.139,73.2133,7.8,7.1,2.6,8,0.62,1.8,88.37
-"Sea Water, Gateway",18.9215,72.8337,8.5,5.3,10,104,1.5,84,57.24
-"Mahim creek, Bandra",19.026,72.8338,8,5,8,56,1.74,210,58.54
-"Bhatsa River, Pise dam",19.3145,73.184,7.8,6.8,3.2,20,1.47,1.8,88.46
-"Vashi Creek, Airoli",19.1601,72.9817,7.6,5.4,5.6,40,0.78,46,71.50
-"Thane Creek, Elephanta",18.9641,72.9428,8.1,4,14,96,1.42,46,53.61
-Malabar Hills,18.9504,72.8038,8.1,4.2,15,92,1.33,84,52.54
-"Versova  beach, andheri",19.1239,72.8168,7.9,4.4,14,216,1.27,94,54.84
-Juhu beach,19.1013,72.8233,8.6,4.8,9,88,1.27,110,54.25
-"Ulhas river, Mohane, Kalyan",19.2559,73.1789,7.6,7.1,2.4,16,0.71,1.8,89.95
-"Ulhas river, Jambhul, Ambernath",19.1994,73.2467,7.6,7,2.6,20,0.58,1.8,89.98
-"Ulhas creek, Mumbra ",19.2013,73.0237,7.6,6.5,4.8,20,1.1,94,76.39
-"Bassein creek, Vasai",19.3195,72.8459,7.8,4.8,11,72,0.88,94,58.43
